Updated the include file and updating the functions to be work or to preven bugs
- The include file now has a comment for what each function does.
- Updated the strlen function with a controll if the text is NULL. Added
better controll to the create string function with better string
capacity allocation, easier to understand copy of the input string to
the dstring with memcpy.
- Improoved readibility for some functions.
- For the append and preppend fucntions added better controll and
allocation for the new lenght. And used memcpy for an easier copy of the
string. And used memmove to shift the original string by n bytes to make
space to the string to preppend.