Updated some other functions for better readibility and better logic
- Updated the include for the string_cmp and the print_string functions
so they use a single pointer since wont modify anythign.
- Updated the substring function with better border case checks, using a
dstring and not a char * for the temp buffer. Checks if the dimension of
the temp buff is enought if not it encreases the dimensions and
reallocates the temp buff. And uses memcpy to copy the substring.
- In string_cmp used a single pointer and just made the logic more
readeble but nothing changed.
- In print string added just a return if the check gives NULL.